# Dispatch `iloc` calls to `getitem`