--- a/dask/dataframe/core.py
+++ b/dask/dataframe/core.py
@@ -3482,6 +3482,7 @@ class DataFrame(_Frame):
         """
         from .indexing import _iLocIndexer
 
+        # For dataframes with unique column names, this will be transformed into a __getitem__ call
         return _iLocIndexer(self)
 
     def __len__(self):