--- a/dask/dataframe/indexing.py
+++ b/dask/dataframe/indexing.py
@@ -58,7 +58,16 @@ class _iLocIndexer(_IndexerBase):
         if iindexer != slice(None):
             raise NotImplementedError(msg)
 
-        return self._iloc(iindexer, cindexer)
+        if len(set(self.obj.columns)) < len(self.obj.columns):
+            # if there are any duplicate column names, do an iloc
+            return self._iloc(iindexer, cindexer)
+        else:
+            # otherwise dispatch to dask.dataframe.core.DataFrame.__getitem__
+            return self._getitem(cindexer)
+
+    def _getitem(self, cindexer):
+        col_names = self.obj.columns[cindexer]
+        return self.obj.__getitem__(col_names)
 
     def _iloc(self, iindexer, cindexer):
         assert iindexer == slice(None)