--- a/dask/dataframe/indexing.py
+++ b/dask/dataframe/indexing.py
@@ -63,11 +63,8 @@ class _iLocIndexer(_IndexerBase):
             return self._iloc(iindexer, cindexer)
         else:
             # otherwise dispatch to dask.dataframe.core.DataFrame.__getitem__
-            return self._getitem(cindexer)
-
-    def _getitem(self, cindexer):
-        col_names = self.obj.columns[cindexer]
-        return self.obj.__getitem__(col_names)
+            col_names = self.obj.columns[cindexer]
+            return self.obj.__getitem__(col_names)
 
     def _iloc(self, iindexer, cindexer):
         assert iindexer == slice(None)