--- a/dask/dataframe/tests/test_indexing.py
+++ b/dask/dataframe/tests/test_indexing.py
@@ -552,14 +552,27 @@ def test_iloc_raises():
         ddf.iloc[:, [5, 6]]
 
 
-def test_iloc_dup_column_no_getitem():
+def test_iloc_duplicate_columns():
     df = pd.DataFrame({"A": [1, 2], "B": [3, 4], "C": [5, 6]})
     ddf = dd.from_pandas(df, 2)
+    df.columns = ["A", "A", "C"]
     ddf.columns = ["A", "A", "C"]
 
     selection = ddf.iloc[:, 2]
-
+    # Check that `iloc` is called instead of getitem
     assert any([key.startswith("iloc") for key in selection.dask.layers.keys()])
+
+    select_first = ddf.iloc[:, 1]
+    assert_eq(select_first, df.iloc[:, 1])
+
+    select_zeroth = ddf.iloc[:, 0]
+    assert_eq(select_zeroth, df.iloc[:, 0])
+
+    select_list_cols = ddf.iloc[:, [0, 2]]
+    assert_eq(select_list_cols, df.iloc[:, [0, 2]])
+
+    select_negative = ddf.iloc[:, -1:-3:-1]
+    assert_eq(select_negative, df.iloc[:, -1:-3:-1])
 
 
 def test_iloc_dispatch_to_getitem():
@@ -570,6 +583,18 @@ def test_iloc_dispatch_to_getitem():
 
     assert all([not key.startswith("iloc") for key in selection.dask.layers.keys()])
     assert any([key.startswith("getitem") for key in selection.dask.layers.keys()])
+
+    select_first = ddf.iloc[:, 1]
+    assert_eq(select_first, df.iloc[:, 1])
+
+    select_zeroth = ddf.iloc[:, 0]
+    assert_eq(select_zeroth, df.iloc[:, 0])
+
+    select_list_cols = ddf.iloc[:, [0, 2]]
+    assert_eq(select_list_cols, df.iloc[:, [0, 2]])
+
+    select_negative = ddf.iloc[:, -1:-3:-1]
+    assert_eq(select_negative, df.iloc[:, -1:-3:-1])
 
 
 def test_iloc_out_of_order_selection():