- [x] Tests added / passed
- [x] Passes `black dask` / `flake8 dask`

Alternate solution to speeding up certain `iloc` calls xref #6345 #6264 

This is a much smaller change and also provides a nice speedup for my simple local benchmarks:
timeseries parquet file created with dask.dataframe.demo with 4 columns and 34447680 rows

current master:

```
In [3]: %time len(ddf)
CPU times: user 11.8 s, sys: 1.89 s, total: 13.7 s
Wall time: 7.1 s
Out[3]: 34447680

In [4]: %time len(ddf)
CPU times: user 10.1 s, sys: 1.6 s, total: 11.7 s
Wall time: 5.2 s
Out[4]: 34447680
```

this branch:

```
In [3]: %time len(ddf)
CPU times: user 5.9 s, sys: 777 ms, total: 6.68 s
Wall time: 3.79 s
Out[3]: 34447680

In [4]: %time len(ddf)
CPU times: user 4.01 s, sys: 504 ms, total: 4.51 s
Wall time: 1.72 s
Out[4]: 34447680
```