@mrocklin 2020-06-29
This PR may win the compute-hours-saved / lines-of-code contest