@gforsyth 2020-06-30
> Can you add a couple sanity tests that the result of a `df.iloc[:, ...]` is correct with duplicate columns?
> 
> * `df.iloc[:, 0]` with duplicate columns and different values for the columns with the same name
> * `df.iloc[:, list_of_integers]` selecting some duplicates and some non-duplicate columns
> * `df.iloc[:, negative_slice]`

Added these tests with duplicate columns and without.